Allow test suite to work against live Redis
While we don't usually want a test quite to depend against an external
service, it can be reassuring to occasionally test low-level libraries
which depend heavily on external service behaviour against those
services in real time.  To this end, we make the default test suite
`mix test` run only the portions of the test suite that can be run
without first configuring Redis.  Then, we modify configurations so
that the test application wires up the in-memory mock for tests that
don't require Redis.  Finally, we write a basic test using Redis in
real time and Tasks to verify that work gets done in proper order.